ai-service/requirements.txt
# requirements.txt - Fixed version
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI Integration
openai==1.3.7
tiktoken==0.5.2

# Database & Vector Store (Fixed: use asyncpg instead of psycopg2)
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# Caching & Queue
redis==5.0.1

# HTTP & Security
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Monitoring & Logging
structlog==23.2.0

# ML & Data Processing
numpy==1.24.4
pandas==2.1.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1